requirements.txt
python-dotenv
yandex-music
telethon
mutagen
winsdk
